% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{aca_citation_format}
\alias{aca_citation_format}
\title{Add Citation Style Language (CSL) file to directory}
\usage{
aca_citation_format(citation_style = "nature", dir = "report")
}
\arguments{
\item{citation_style}{Name of the CSL you wish to use. Taken from
https://github.com/citation-style-language/styles}

\item{dir}{Directory to add CSL file to}
}
\value{

}
\description{
Add Citation Style Language (CSL) file to directory
}
