% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{aca_setup_paper}
\alias{aca_setup_paper}
\title{Set-up repo for paper}
\usage{
aca_setup_paper(
  paper_name = "manuscript",
  paper_type = "normal",
  csl = "nature"
)
}
\arguments{
\item{paper_name}{Document name}

\item{paper_type}{Future proofing for different paper types}

\item{csl}{Type of Citation Style Language you wish to use}
}
\value{

}
\description{
Set-up repo for paper
}
