% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{aca_simple_table}
\alias{aca_simple_table}
\title{Create a formatted table using common defaults}
\usage{
aca_simple_table(data, fontsize = 7, autofit = TRUE, bold_first_column = TRUE)
}
\arguments{
\item{data}{Data to present}

\item{fontsize}{Fontsize}

\item{autofit}{Autofit?}

\item{bold_first_column}{Whether or not to bold first column.}
}
\value{

}
\description{
Create a formatted table using common defaults
}
