% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{aca_use_github}
\alias{aca_use_github}
\title{Add GitHub package to 0_library.R}
\usage{
aca_use_github(user, package)
}
\arguments{
\item{user}{GitHub username}

\item{package}{Package name}
}
\value{

}
\description{
Add GitHub package to 0_library.R
}
