% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add location.R
\name{add_locations}
\alias{add_locations}
\title{Add Lon and Lat for Places}
\usage{
add_locations(df, df_loc = loc_df)
}
\arguments{
\item{df}{Text df.}

\item{df_loc}{Loc df, defaults to in house}
}
\value{
Returns data frame of locations names and lon and lat.
}
\description{
Join from in house data.
}
