% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add speaker.R, R/clean words.R
\name{clean_text}
\alias{clean_text}
\alias{add_speaker}
\alias{clean_text}
\title{Add speaker categorical variable}
\usage{
add_speaker(df, keep = T)

clean_text(df, var.)
}
\arguments{
\item{df}{Takes df with sentence level text.}

\item{keep}{Defaults to TRUE, if FALSE filters out "jorn"...}

\item{df}{Takes df with}
}
\value{
Returns data frame

Returns data frame
}
\description{
Mutates variable \code{"speaker."} to either \code{"jorn"} or \code{"pm"}.

Removes numeric characters, stop words and white space from words...
}
