% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch words.R
\name{fetch_sentences}
\alias{fetch_sentences}
\title{Get data for MT speeches as sentences}
\usage{

  fetch_sentences(path = "/TrendLock/showcase projects/turnbull/media_sentences.feather",
  sample = T)
}
\arguments{
\item{path}{Defaults to storage}
}
\value{
Returns data frame
}
\description{
"Anyone who thinks it's smart to cut immigration is sentencing Australia to poverty" - MT
}
