% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sent to word.R
\name{sent_to_words}
\alias{sent_to_words}
\title{Convert sentence stings to words}
\usage{
sent_to_words(df, remove_jorn = T)
}
\arguments{
\item{df}{Takes df with}

\item{remove_jorn}{Defaults to TRUE, filters out "jorn" speaker.}
}
\value{
Returns data frame
}
\description{
Converts strings that are sentences to one word per row.
Uses \code{tidttext} function \code{unnest_tokens()}. Brilliant!
}
